cover_screen/client/python/pyproject.toml
[project]
name = "cover-screen-client"
version = "0.1.0"
description = "Cover Screen Client"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pyzmq",
  "pillow",
  "psutil",
  "pyyaml",
  "stransi",
  "rich",
  "ochre",
]
